+++ src/components/AsideImage/AsideImage.css
@@ -5,7 +5,7 @@ aside {
 img {
     object-fit: cover;
     max-width: 100%;
-    background: url('../img/preloader.gif') no-repeat center;
+    background: url('../../img/preloader.gif') no-repeat center;
 }
 
 figcaption {
