+++ src/components/AsideImage/AsideImage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BASE_URL } from '../constants';
+import { BASE_URL } from '../../constants';
 import './AsideImage.css';
 
 interface IProps {
